code cleanup - fixing compiler warnings

+++ lib/logstreamxx/logstreambuf.cpp
@@ -65,7 +65,7 @@ namespace logstreamxx {
 		int n = strftime( buffer, 16, "%b %e %T", ti );
 
 		// append milliseconds
-		sprintf( ( buffer + n ), ".%ld", tv.tv_usec );
+		sprintf( ( buffer + n ), ".%ld", (long int) tv.tv_usec );
 
 		return buffer;
